backend/pyproject.toml
[project]
name = "via-backend"
version = "0.1.0"
description = "Backend service for the Video Intelligence Analyzer (VIA)."
authors = [{name = "Naseef", email = "nasifrks@hotmail.com"}]
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlmodel",
    "pydantic-settings",
    "python-multipart",
    "requests",
    "aiofiles",
    "orjson",
]

[project.optional-dependencies]
full = [
    "openai>=1.0.0",
    "ultralytics>=8.0.0",
    "whisper @ git+https://github.com/openai/whisper.git",
    "pytube",
    "python-dotenv",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
